+++ CHANGES.rst
@@ -8,6 +8,11 @@ details, see the commit logs at https://github.com/girder/slicer_package_manager
 Next Release
 ============
 
+Tests
+-----
+
+* Improve PackageMetadataChecksumUpdate tests to check extensions package.
+
 0.8.0
 =====